requirements.txt
flask
Pillow
git+https://github.com/huggingface/diffusers
torch
transformers
accelerate
xformers==0.0.17
opencv-python
git+https://github.com/facebookresearch/segment-anything.git
replicate